+++ packages/scenes/src/variables/adhoc/AdHocFiltersVariable.tsx
@@ -49,11 +49,12 @@ export class AdHocFiltersVariable
       this._subs.add(
         this.state.set.subscribeToState((newState, prevState) => {
           if (newState.filters !== prevState.filters) {
-            this._filtersChanged(newState.filters);
+            this._updateFilterExpression(newState.filters, true);
           }
         })
       );
-      this._filtersChanged(this.state.set.state.filters);
+
+      this._updateFilterExpression(this.state.set.state.filters, false);
     });
   }
 
@@ -61,15 +62,25 @@ export class AdHocFiltersVariable
     return this.state.filterExpression;
   }
 
-  private _filtersChanged(filters: AdHocVariableFilter[]) {
+  private _updateFilterExpression(filters: AdHocVariableFilter[], publishEvent: boolean) {
     let expr = '';
 
     for (const filter of filters) {
       expr += `${this._renderFilter(filter)},`;
     }
 
+    if (expr.length > 0) {
+      expr = expr.slice(0, -1);
+    }
+
+    if (expr === this.state.filterExpression) {
+      return;
+    }
+
     this.setState({ filterExpression: expr });
-    this.publishEvent(new SceneVariableValueChangedEvent(this), true);
+    if (publishEvent) {
+      this.publishEvent(new SceneVariableValueChangedEvent(this), true);
+    }
   }
 
   private _renderFilter(filter: AdHocVariableFilter) {