+++ packages/scenes/src/variables/variants/IntervalVariable.tsx
@@ -57,7 +57,12 @@ export class IntervalVariable
     const update: Partial<IntervalVariableState> = {};
     const val = values[this.getKey()];
     if (typeof val === 'string') {
-      update.value = val;
+      // support old auto interval url value
+      if (val.startsWith('$__auto_interval_')) {
+        update.value = AUTO_VARIABLE_VALUE;
+      } else {
+        update.value = val;
+      }
     }
     this.setState(update);
   }